requirements.txt
biopython
numpy
memoized
